refactor: change all filename with track keyword to trace
- fix warning unused propse
