+++ src/@types/navigations/scenes/traces.d.ts
@@ -2,12 +2,9 @@ import type { NativeStackScreenProps } from "@react-navigation/native-stack";
 
 import { SCREEN_NAMES } from "@services/constants/screen";
 
-type TraceListSceneStackParamList = {
+type TracesSceneStackParamList = {
   [SCREEN_NAMES.TRACE_LIST]: undefined;
   [SCREEN_NAMES.TRACE_DETAIL]: { id: string };
 };
 
-type TraceListSceneStackProps<ScreenName> = NativeStackScreenProps<
-  TraceListSceneStackParamList,
-  ScreenName
->;
+type TracesSceneProps<ScreenName> = NativeStackScreenProps<TracesSceneStackParamList, ScreenName>;
