+++ src/navigations/app.tsx
@@ -8,7 +8,7 @@ import AccountScreen from "@screens/AccountScreen";
 import TraceCreateScreen from "@screens/TraceCreateScreen/TraceCreateScreen";
 import { CUSTOM_HEADER_TITLE, SCREEN_NAMES } from "@services/constants/screen";
 
-import TraceListSceneNavigation from "./scenes/trackList";
+import TraceListSceneNavigation from "./scenes/traces";
 
 const AppStack = createBottomTabNavigator<AppStackParamList>();
