+++ src/screens/TracesScene/TraceListScreen/TraceListScreen.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 
-import { TraceListSceneStackProps } from "@@types/navigations/scenes/trackList";
+import { TracesSceneProps } from "@@types/navigations/scenes/traces";
 import { HeadingText } from "@components/Texts";
 import { SCREEN_NAMES } from "@services/constants/screen";
 
-const TraceListScreen = ({}: TraceListSceneStackProps<SCREEN_NAMES.TRACE_LIST>) => {
+const TraceListScreen = ({}: TracesSceneProps<SCREEN_NAMES.TRACE_LIST>) => {
   return <HeadingText type="h1">TraceListScreen</HeadingText>;
 };
